Feature: - Timer records sessions to DB

+++ src/focusseeds/clock.py
@@ -3,15 +3,26 @@ from textual import on
 from textual.widgets import Button, Static, Label, Input
 from textual.containers import Horizontal, Vertical
 
-from focusseeds.numbers import NUMBERS_DICT
+import pygame
+
+from focusseeds._numbers import NUMBERS_DICT
 from focusseeds.validators import ValueFrom5to300
 from focusseeds.confirmation_popup import ConfirmPopup
+from focusseeds.db import DatabaseManager
+from focusseeds.config import UNFS_BRAAM
 
 
 class Clock(Static):
+    """
+    Clock widget functionalities:
+        - Time display for both timer and stopwatch modes.
+        - Input field to set the timer duration.
+        - Button to start, stop, or cancel the timer/stopwatch.
+        - Play sound when session is successful
+    """
+
     DEFAULT_CSS = """
     Clock {
-        border: heavy yellow;
         width: 100%;
         height: 100%;
     }
@@ -20,7 +31,7 @@ class Clock(Static):
         align: center middle;
         height: 60%;
     }
-    
+
     #seed-wrapper {
         align: center middle;
         height: 40%;
@@ -55,18 +66,28 @@ class Clock(Static):
             value='45',
             placeholder='Type 5 to 300',
             restrict=r'^\d{1,3}$',
-            validators=[ValueFrom5to300()],
+            # validators=[ValueFrom5to300()],
             id='session-duration'
         )
-        # Data
+        # Mode
+        self.active_session = False
         self._clock_mode: str = 'Timer'
-        self._stop_watch_started: datetime = datetime.now()
-        self._timer_length: int = 0
-        self._cancel_timer_counter_default = 3
-        self._cancel_timer_counter: int = 0
-
+        # Stopwatch
+        self._stopwatch_started: datetime = datetime.now()
+        # Timer - sessions are stored in seconds
+        self._timer_session_len: int = 0
+        self._timer_remaining_session: int = 0
+        # Cancel session
+        self._cancel_session_counter_default = 60
+        self._cancel_session_counter: int = 0
+        # Clock - When ever user start clock they get restarted
+        # Used to print time on a screen
+        self._minutes: float = 0
+        self._seconds: float = 0
         # Intervals
         self._intervals = []
+        # External classes
+        self.db = DatabaseManager()
 
     def compose(self):
         with Horizontal(id='clock-wrapper'):
@@ -86,22 +107,77 @@ class Clock(Static):
             yield self._seed_button
 
     @on(Button.Pressed, '#seed-bt')
-    def _start_timer(self) -> None:
-        """Start timer or stopwatch depend on the chosen mode"""
+    def _seed_button_clicked(self) -> None:
+        """Start, Cancel, Kill timer or stopwatch depend on the chosen mode"""
+        # Start timer
         if self._seed_button.variant == 'success':
             self._input_filed.visible = False
-            self._start_seeding()
+            self._start_session()
+            self.active_session = True
+        # Cancel timer
         elif self._seed_button.variant == 'warning':
             self._reset_clock()
+        # Kill/End timer
         else:
-            popup = ConfirmPopup(message='Do you want to kill the session?')
-            self.app.push_screen(popup, self._record_killed_session)
+            # Timer
+            if self._clock_mode == 'Timer':
+                # If user accepts to kill session, `self._not_successful_session` is called
+                popup = ConfirmPopup(message='Do you want to kill the session?')
+                self.app.push_screen(popup, self._not_successful_session)
+            # Stopwatch
+            else:
+                self._successful_session(self._minutes)
 
-    def _record_killed_session(self, boolean: bool) -> None:
+    def _start_session(self) -> None:
+        """Start a timer or stopwatch session based on the current clock mode."""
+        # Initialize cancel timer counter
+        self._cancel_session_counter = self._cancel_session_counter_default
+
+        if self._clock_mode == 'Timer':
+            self._timer_session_len = int(self.query_one(Input).value) * 60
+            self._timer_remaining_session = self._timer_session_len
+        else:
+            self._stopwatch_started = datetime.now()
+
+        # Select appropriate clock mode function
+        clock_mode_func = self._timer if self._clock_mode == 'Timer' else self._stopwatch
+
+        # Set intervals for updating clock and managing cancel timer
+        session_interval = self.set_interval(1, clock_mode_func)
+        cancel_session_interval = self.set_interval(1, self._cancel_session)
+        self._intervals.extend((session_interval, cancel_session_interval))
+
+        # Set button variant to 'warning' to indicate session is ongoing
+        self._seed_button.variant = 'warning'
+
+    def _stopwatch(self) -> None:
+        """Update variable used by stopwatch and update displayed time"""
+        elapsed_time = datetime.now() - self._stopwatch_started
+        self._minutes, self._seconds = divmod(elapsed_time.total_seconds(), 60)
+        self._update_clock()
+
+    def _timer(self) -> None:
+        """Update variable used by timer and update displayed time"""
+        self._timer_remaining_session -= 1
+        # When end of the session
+        if self._timer_remaining_session == 0:
+            self._successful_session(self._timer_session_len // 60)
+        else:
+            self._minutes, self._seconds = divmod(self._timer_remaining_session, 60)
+            self._update_clock()
+
+    def _successful_session(self, minutes: float) -> None:
+        """Play song, add successful session to DB and reset clock"""
+        pygame.mixer.music.load(UNFS_BRAAM)
+        pygame.mixer.music.play()
+        self.db.create_session_entry(int(minutes), 1)
+        self._reset_clock()
+
+    def _not_successful_session(self, boolean: bool) -> None:
         """Add killed session to DB and reset clock"""
         if not boolean:
             return
-        # TODO: Add record to DB
+        self.db.create_session_entry(int(self._minutes), 0)
         self._reset_clock()
 
     def _reset_clock(self) -> None:
@@ -117,53 +193,24 @@ class Clock(Static):
         self._seed_button.label = 'Seed'
         # Unhidden Input
         self._input_filed.visible = True
+        # Session Variable
+        self.active_session = False
         # Stop intervals
-        self._stop_intervals()
-
-    def _stop_intervals(self):
-        """Stop all active intervals."""
         for interval in self._intervals:
             interval.stop()
         self._intervals.clear()
 
-    def _start_seeding(self) -> None:
-        """Start clock"""
-        self._cancel_timer_counter = self._cancel_timer_counter_default
-        if self._clock_mode == 'Timer':
-            self._timer_length = int(self.query_one(Input).value) * 60
-            interval1 = self.set_interval(1, self._timer)
-        else:
-            self._stop_watch_started = datetime.now()  # Reload time
-            interval1 = self.set_interval(1, self._stopwatch)
-
-        self._seed_button.variant = 'warning'
-        interval2 = self.set_interval(1, self._start_cancel_timer, repeat=30)
-        self._intervals.extend((interval1, interval2))
-
-    def _stopwatch(self) -> None:
-        """Update variable used by stopwatch and update displayed time"""
-        elapsed_time = datetime.now() - self._stop_watch_started
-        minutes, seconds = divmod(elapsed_time.total_seconds(), 60)
-        self._update_clock(minutes, seconds)
-
-    def _timer(self) -> None:
-        """Update variable used by timer and update displayed time"""
-        self._timer_length -= 1
-        minutes, seconds = divmod(self._timer_length, 60)
-        self._update_clock(minutes, seconds)
-
-    def _update_clock(self, minutes, seconds):
+    def _update_clock(self):
         """Update displayed digits"""
-        minutes_str = str(int(minutes)).zfill(2)
-        seconds_str = str(int(seconds)).zfill(2)
-        print(int(minutes))
+        minutes_str = str(int(self._minutes)).zfill(2)
+        seconds_str = str(int(self._seconds)).zfill(2)
 
-        if int(minutes) >= 100:
+        if int(self._minutes) >= 100:
             self._h_min.update(NUMBERS_DICT[minutes_str[-3]])
         else:
             self._h_min.update('')
 
-        if int(minutes) >= 10:
+        if int(self._minutes) >= 10:
             self._t_min.update(NUMBERS_DICT[minutes_str[-2]])
         else:
             self._t_min.update('')
@@ -172,17 +219,21 @@ class Clock(Static):
         self._t_sec.update(NUMBERS_DICT[seconds_str[-2]])
         self._u_sec.update(NUMBERS_DICT[seconds_str[-1]])
 
-    def _start_cancel_timer(self):
-        """Allow user to cancel timer in first 30 seconds"""
-        self._cancel_timer_counter -= 1
-        if self._cancel_timer_counter > 0:
-            self._seed_button.label = f'Cancel ({self._cancel_timer_counter})'
+    def _cancel_session(self):
+        """
+        Allow user to cancel timer in first
+        `self._cancel_timer_counter_default` seconds
+        """
+        self._cancel_session_counter -= 1
+        if self._cancel_session_counter > 0:
+            self._seed_button.label = f'Cancel ({self._cancel_session_counter})'
         else:
-            self._seed_button.label = 'Kill'
+            # Set button when the cancel time has ended
+            self._seed_button.label = 'Kill' if self._clock_mode == 'Timer' else 'End'
             self._seed_button.variant = 'error'
 
     @on(Input.Changed, '#session-duration')
-    def _session_duration_changed(self, event: Input.Changed) -> None:
+    def _timer_session_duration_changed(self, event: Input.Changed) -> None:
         """If the session duration is not correct block start button."""
         if event.input.is_valid:
             self.query_one('#seed-bt', Button).disabled = False
@@ -190,20 +241,21 @@ class Clock(Static):
             self.query_one('#seed-bt', Button).disabled = True
 
     def change_clock_mode(self):
-        """Change between `timer` and `stopwatch`"""
+        """
+        Change between `timer` and `stopwatch`
+        Remove input when changing to stopwatch
+        Insert input when changing to timer
+        """
         if self._clock_mode == 'Timer':
             self._clock_mode = self.app.title = 'Stopwatch'
             self.query_one(Input).remove()
         else:
             self._clock_mode = self.app.title = 'Timer'
-            self.mount(
-                Input(
-                    value='45',
-                    placeholder='Type number from 5 to 300',
-                    restrict=r'^\d{1,3}$',
-                    validators=[ValueFrom5to300()],
-                    id='session-duration'
-                ),
-                before='#seed-bt',
+            self._input_filed = Input(
+                value='45',
+                placeholder='Type number from 5 to 300',
+                restrict=r'^\d{1,3}$',
+                validators=[ValueFrom5to300()],
+                id='session-duration'
             )
-
+            self.mount(self._input_filed, before='#seed-bt')
